+++ packages/saved-views-client/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased](https://github.com/iTwin/saved-views/tree/HEAD/packages/saved-views-client)
 
+## [0.4.0](https://github.com/iTwin/saved-views/tree/client-v0.4.0/packages/saved-views-client/) - 2024-07-22
+
 ### Breaking changes
 
 * `SavedViewsClient` interface changes